Don't remove comments so downstream users get docs
Causing TSC to remove comments means that downstream users don't get
access to docs when they use a language server. This commit will cause
TSC to emit these comments when it generates `.d.ts` files and `.js `
files.

This results in bigger binaries, but this is not expected to be a
problem.